Removed ImageFileExtensionMap class
